🧬 Migrate the data model about *mocked_apis* settings back to *__init__* module.
