+++ pymock_api/model/api_config/__init__.py
@@ -2,7 +2,7 @@
 
 content ...
 """
-from typing import Any, Dict, Optional, Union
+from typing import Any, Dict, List, Optional, Union
 
 from ..._utils import YAML
 from ..._utils.file_opt import _BaseFileOperation
@@ -13,12 +13,222 @@ from .apis import (
     HTTPRequest,
     HTTPResponse,
     MockAPI,
-    MockAPIs,
     ResponseProperty,
 )
 from .base import BaseConfig
 from .item import IteratorItem
 from .template import TemplateConfig
+
+
+class MockAPIs(_Config):
+    """*The **mocked_apis** section*"""
+
+    _template: TemplateConfig
+    _base: Optional[BaseConfig]
+    _apis: Dict[str, Optional[MockAPI]]
+
+    _need_template_in_config: bool = True
+
+    def __init__(
+        self,
+        template: Optional[TemplateConfig] = None,
+        base: Optional[BaseConfig] = None,
+        apis: Dict[str, Optional[MockAPI]] = {},
+    ):
+        self._template = TemplateConfig() if template is None else template
+        self._base = base
+        self._apis = apis
+
+    def __len__(self):
+        return len(self.apis.keys())
+
+    def _compare(self, other: "MockAPIs") -> bool:
+        return self.base == other.base and self.apis == other.apis
+
+    @property
+    def template(self) -> TemplateConfig:
+        return self._template
+
+    @template.setter
+    def template(self, template: Union[dict, TemplateConfig]) -> None:
+        if template:
+            if isinstance(template, dict):
+                self._template = TemplateConfig().deserialize(data=template)
+            elif isinstance(template, TemplateConfig):
+                self._template = template
+            else:
+                raise TypeError("Setter *MockAPIs.template* only accepts dict or TemplateConfig type object.")
+
+    @property
+    def base(self) -> Optional[BaseConfig]:
+        return self._base
+
+    @base.setter
+    def base(self, base: Union[dict, BaseConfig]) -> None:
+        if base:
+            if isinstance(base, dict):
+                self._base = BaseConfig().deserialize(data=base)
+            elif isinstance(base, BaseConfig):
+                self._base = base
+            else:
+                raise TypeError("Setter *MockAPIs.base* only accepts dict or BaseConfig type object.")
+        else:
+            self._base = None
+
+    @property
+    def apis(self) -> Dict[str, Optional[MockAPI]]:
+        return self._apis
+
+    @apis.setter
+    def apis(self, apis: Dict[str, Union[dict, MockAPI]]) -> None:
+        if apis:
+            if not isinstance(apis, dict):
+                raise TypeError("Setter *MockAPIs.apis* only accepts dict or MockAPI type object.")
+
+            ele_types = set(list(map(lambda v: isinstance(v, MockAPI), apis.values())))
+            if len(ele_types) != 1:
+                raise ValueError("It has multiple types of the data content. Please unify these objects data type.")
+
+            if False in ele_types:
+                self._apis = {}
+                for api_name, api_config in apis.items():
+                    self._apis[api_name] = MockAPI().deserialize(data=(api_config or {}))
+            else:
+                self._apis = apis  # type: ignore
+        else:
+            self._apis = {}
+
+    @property
+    def set_template_in_config(self) -> bool:
+        return self._need_template_in_config
+
+    @set_template_in_config.setter
+    def set_template_in_config(self, _set: bool) -> None:
+        self._need_template_in_config = _set
+
+    def serialize(self, data: Optional["MockAPIs"] = None) -> Optional[Dict[str, Any]]:
+        template = (data.template if data else None) or self.template
+        base = (data.base if data else None) or self.base
+        apis = (data.apis if data else None) or self.apis
+        if not (base and apis):
+            return None
+
+        # Process section *base*
+        api_info = {  # type: ignore[var-annotated]
+            "base": BaseConfig().serialize(data=base),
+            "apis": {},
+        }
+        if self._need_template_in_config:
+            api_info["template"] = template.serialize()
+
+        # Process section *apis*
+        all_mocked_apis = {}
+        for api_name, api_config in apis.items():
+            all_mocked_apis[api_name] = MockAPI().serialize(data=api_config)
+        api_info["apis"] = all_mocked_apis
+
+        return api_info
+
+    @_Config._ensure_process_with_not_empty_value
+    def deserialize(self, data: Dict[str, Any]) -> Optional["MockAPIs"]:
+        """Convert data to **MockAPIs** type object.
+
+        The data structure should be like following:
+
+        * Example data:
+        .. code-block:: python
+
+            {
+                'base': {'url': '/test/v1'},
+                'apis': {
+                    'google_home': {
+                        'url': '/google',
+                        'http': {
+                            'request': {
+                                'method': 'GET',
+                                'parameters': [{'param1': 'val1'}]
+                            },
+                            'response': {
+                                'value': 'This is Google home API.'
+                            }
+                        }
+                    },
+                    'test_home': {
+                        'url': '/google',
+                        'http': {
+                            'request': {
+                                'method': 'GET',
+                                'parameters': [{'param1': 'val1'}]
+                            },
+                            'response': {
+                                'value': '{ "responseCode": "200", "errorMessage": "OK", "content": "This is Test home." }'
+                            }
+                        },
+                        'cookie': [{'TEST': 'cookie_value'}]
+                    },
+                    'youtube_home': {
+                        'url': '/youtube',
+                        'http': {
+                            'request': {
+                                'method': 'GET',
+                                'parameters': [{'param1': 'val1'}]
+                            },
+                            'response': {'value': 'youtube.json'}
+                        },
+                        'cookie': [{'USERNAME': 'test'}, {'SESSION_EXPIRED': '2023-12-31T00:00:00.000'}]
+                    }
+                }
+            }
+
+        Args:
+            data (Dict[str, Any]): Target data to convert.
+
+        Returns:
+            A **MockAPIs** type object.
+
+        """
+        # Processing section *template*
+        template_info = data.get("template", {})
+        if not template_info:
+            self._need_template_in_config = False
+        self.template = TemplateConfig().deserialize(data=template_info)
+
+        # Processing section *base*
+        base_info = data.get("base", None)
+        self.base = BaseConfig().deserialize(data=base_info)
+
+        # Processing section *apis*
+        mocked_apis_info = data.get("apis", {})
+        self.apis = {}
+        if mocked_apis_info:
+            for mock_api_name in mocked_apis_info.keys():
+                self.apis[mock_api_name] = MockAPI().deserialize(data=mocked_apis_info.get(mock_api_name, None))
+        return self
+
+    def get_api_config_by_url(self, url: str, base: Optional[BaseConfig] = None) -> Optional[MockAPI]:
+        url = url.replace(base.url, "") if base else url
+        for k, v in self._apis.items():
+            if v and v.url == url:
+                return self._apis[k]
+        return None
+
+    def get_all_api_config_by_url(self, url: str, base: Optional[BaseConfig] = None) -> Dict[str, MockAPI]:
+        url = url.replace(base.url, "") if base else url
+        all_api_configs: Dict[str, MockAPI] = {}
+        for k, v in self._apis.items():
+            if v and v.url == url:
+                all_api_configs[v.http.request.method.upper()] = self._apis[k]  # type: ignore[union-attr,assignment]
+        return all_api_configs
+
+    def group_by_url(self) -> Dict[str, List[MockAPI]]:
+        apis = self._apis
+        aggregated_apis: Dict[str, List[MockAPI]] = {}
+        for api_name, api_config in apis.items():
+            assert api_config and api_config.url
+            one_url_details = aggregated_apis.get(api_config.url, [])
+            one_url_details.append(api_config)
+            aggregated_apis[api_config.url] = one_url_details
+        return aggregated_apis
 
 
 class APIConfig(_Config):
